+++ CHANGELOG.md
@@ -44,7 +44,7 @@
 * [#271](https://github.com/bbatsov/rubocop/issues/271) - always allow line breaks inside hash literal braces
 * [#270](https://github.com/bbatsov/rubocop/issues/270) - fixed a false positive in ParenthesesAroundCondition
 * [#288](https://github.com/bbatsov/rubocop/issues/288) - get config parameter AllCops/Excludes from highest config file in path
-
+* [#276](https://github.com/bbatsov/rubocop/issues/276) - let columns start at 1 instead of 0 in all output of column numbers
 
 ## 0.8.2 (06/05/2013)